Consolidated utils location
