+++ client/checkout/classic/index.js
@@ -5,7 +5,7 @@
  */
 import './style.scss';
 import { PAYMENT_METHOD_NAME } from '../constants.js';
-import { getConfig } from '../utils';
+import { getConfig } from 'utils/checkout';
 import WCPayAPI from './../api';
 
 jQuery( function ( $ ) {
