+++ client/data/transactions/reducer.js
@@ -4,7 +4,7 @@
  * Internal dependencies
  */
 import TYPES from './action-types';
-import { getResourceId } from '../util';
+import { getResourceId } from 'utils/data';
 
 const defaultState = { summary: {} };
