+++ client/data/transactions/test/reducer.js
@@ -5,7 +5,7 @@
  */
 import reducer from '../reducer';
 import types from '../action-types';
-import { getResourceId } from '../../util';
+import { getResourceId } from 'utils/data';
 
 describe( 'Transactions reducer tests', () => {
 	const mockQuery = { paged: '2', perPage: '50' };
