+++ client/data/transactions/test/selectors.js
@@ -3,7 +3,7 @@
 /**
  * Internal dependencies
  */
-import { getResourceId } from '../../util';
+import { getResourceId } from 'utils/data';
 import {
 	getTransactions,
 	getTransactionsError,
